It f*cking works (on Linux, Windows is still a nightmare because of SAL annotations)

+++ include/stddef.h
@@ -11,7 +11,7 @@ typedef long long intmax_t;
 /**
  * @brief Used to store the sizes of things.
  */
-typedef unsigned long size_t;
+typedef unsigned long long size_t;
 
 /**
  * @brief wchar_t holds wide characters (i.e. beyond ASCII)
@@ -35,7 +35,7 @@ typedef short wchar_t;
  * it gets the address of the member as if it was
  * part of a structure starting at 0, which means it's
  * just the offset of the member from 0 or in other words the
- * start of the structure.
+ * start of the structure. *taps head to indicate bigbrain*
  */
 #define offsetof(structure, member) \
 	((size_t)&(((structure *)0)->member))
